--- a/inference_lib/src/aqlm/inference.py
+++ b/inference_lib/src/aqlm/inference.py
@@ -92,7 +92,7 @@ class QuantizedLinear(nn.Module):
             get_backward_pass_kernel(self.codebooks, True),
         )
 
-        self.use_gemv_rule = lambda input: math.prod(input.shape[:-1]) <= 32
+        self.use_gemv_rule = lambda input: math.prod(input.shape[:-1]) <= 16
 
 
 def _get_autograd_matmul_op(forward_pass_kernel, backward_pass_kernel):