--- a/inference_lib/src/aqlm/inference_kernels/cuda_kernel.cpp
+++ b/inference_lib/src/aqlm/inference_kernels/cuda_kernel.cpp
@@ -62,6 +62,15 @@ void code2x8_matvec_cuda(
   bool use_bfloat16
 );
 
+void code2x8_dequant_cuda(
+  const void* A,
+        void* C,
+  const void* codebook,
+  int prob_m,
+  int prob_k,
+  bool use_bfloat16
+);
+
 void code1x16_matvec(
   const torch::Tensor& A,
   const torch::Tensor& B,
@@ -220,6 +229,27 @@ torch::Tensor code1x16_matmat_dequant(
   return F::linear(input, weight, bias_2);
 }
 
+torch::Tensor code1x16_matmat_dequant_transposed(
+  const torch::Tensor& input,
+  const torch::Tensor& codes,
+  const torch::Tensor& codebooks,
+  const torch::Tensor& scales,
+  const std::optional<torch::Tensor>& bias
+) {
+  auto weight = code1x16_dequant(
+    codes,
+    codebooks,
+    scales
+  ).transpose(0, 1);
+
+  torch::Tensor bias_2{};
+  if (bias.has_value()) {
+    bias_2 = bias.value();
+  }
+
+  return F::linear(input, weight, bias_2);
+}
+
 void code2x8_matvec(
   const torch::Tensor& A,
   const torch::Tensor& B,
@@ -281,11 +311,83 @@ torch::Tensor code2x8_matmat(
   return output;
 }
 
+torch::Tensor code2x8_dequant(
+  const torch::Tensor& codes,
+  const torch::Tensor& codebooks,
+  const torch::Tensor& scales
+) {
+  auto use_bfloat16 = check_use_bfloat16(codebooks);
+  auto in_features = codes.size(1) * 8;
+  auto out_features = scales.size(0);
+
+  auto weight = torch::zeros({out_features, in_features},
+    torch::TensorOptions()
+      .dtype(codebooks.dtype())
+      .device(codebooks.device())
+  );
+  code2x8_dequant_cuda(
+    codes.data_ptr(),
+    weight.data_ptr(),
+    codebooks.data_ptr(),
+    out_features,
+    in_features,
+    use_bfloat16
+  );
+  weight *= scales.index({"...", 0, 0});
+
+  return weight;
+}
+
+torch::Tensor code2x8_matmat_dequant(
+  const torch::Tensor& input,
+  const torch::Tensor& codes,
+  const torch::Tensor& codebooks,
+  const torch::Tensor& scales,
+  const std::optional<torch::Tensor>& bias
+) {
+  auto weight = code2x8_dequant(
+    codes,
+    codebooks,
+    scales
+  );
+
+  torch::Tensor bias_2{};
+  if (bias.has_value()) {
+    bias_2 = bias.value();
+  }
+
+  return F::linear(input, weight, bias_2);
+}
+
+torch::Tensor code2x8_matmat_dequant_transposed(
+  const torch::Tensor& input,
+  const torch::Tensor& codes,
+  const torch::Tensor& codebooks,
+  const torch::Tensor& scales,
+  const std::optional<torch::Tensor>& bias
+) {
+  auto weight = code2x8_dequant(
+    codes,
+    codebooks,
+    scales
+  ).transpose(0, 1);
+
+  torch::Tensor bias_2{};
+  if (bias.has_value()) {
+    bias_2 = bias.value();
+  }
+
+  return F::linear(input, weight, bias_2);
+}
 
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
   m.def("code1x16_matmat", &code1x16_matmat, "1x16 (2bit) codebook matrix-matrix product through matvec.");
   m.def("code1x16_dequant", &code1x16_dequant, "1x16 (2bit) codebook dequantization.");
   m.def("vllm_dequant", &vllm_dequant, "vllm dequantization.");
   m.def("code1x16_matmat_dequant", &code1x16_matmat_dequant, "1x16 (2bit) codebook matrix-matrix dequantization product.");
+  m.def("code1x16_matmat_dequant_transposed", &code1x16_matmat_dequant_transposed, "1x16 (2bit) codebook matrix-matrix dequantization product for backward pass.");
   m.def("code2x8_matmat", &code2x8_matmat, "2x8 (2bit) codebook matrix-matrix product.");
+  m.def("code2x8_dequant", &code2x8_dequant, "2x8 (2bit) codebook dequantization.");
+  m.def("code2x8_matmat_dequant", &code2x8_matmat_dequant, "2x8 (2bit) codebook matrix-matrix dequantization product.");
+  m.def("code2x8_matmat_dequant_transposed", &code2x8_matmat_dequant_transposed, "2x8 (2bit) codebook matrix-matrix dequantization product for backward pass.");
 }