This PR adds faster dequantization kernels written in CUDA.