@BlackSamorez 2024-03-22
The new dequantization kernels are about 4.5x faster for 1x16, leading to up to 4.5x faster prefill and training.

| num_tokens (batch_size x seq_len) | new gemm, ms/pass | old gemm, ms/pass | gemv, ms/pass | `fp16` baseline |
|--|--|--|--|--|
| 1 | 1.10 | 4.71 | 0.18 | 0.14 |
| 4 | 1.10 | 4.68 | 0.53 | 0.14 |
| 16 | 1.11 | 4.69 | 1.91 | 0.14 |
| 64 | 1.16 | 4.72 | 7.44 | 0.17 |
| 256 | 1.54 | 5.01 | too slow | 0.57 |
| 1024 | 2.42 | 6.12 | too slow | 1.40 |
| 4096 | 6.53 | 10.03 | too slow | 5.31 |
| 16384 | 23.19 | 25.63 | too slow | 21.33 |