@BlackSamorez 2024-03-22
The new dequantization kernels are about 7x faster for 2x8, leading to up to 7x faster prefill and training.
| num_tokens (batch_size x seq_len) | new gemm, ms/pass | old gemm, ms/pass | gemv, ms/pass | `fp16` baseline |
|--|--|--|--|--|
| 1 | 0.79 | 5.44 | 0.13 | 0.14 |
| 4 | 0.77 | 5.44 | 0.27 | 0.14 |
| 16 | 0.79 | 5.46 | 0.87 | 0.14 |
| 64 | 0.84 | 5.50 | 3.27 | 0.17 |
| 256 | 1.22 | 5.93 | too slow | 0.58 |
| 1024 | 2.11 | 6.79 | too slow | 1.41 |
| 4096 | 6.18 | 10.78 | too slow | 5.28 |
| 16384 | 22.68 | 26.64 | too slow | 21.16 |